relation cached. TODO update relation asyn.

+++ app/helpers/relations_helper.rb
@@ -5,7 +5,7 @@ module RelationsHelper
   # 参数
   # relationable 可以发生关系的对象（Project, Publication, User)
   # opts
-  #  - TODO cache 为 true 时，异步 relation_for，页面加载完后才去异步显示
+  #  - TODO cached 为 true 时，异步 relation_for，页面加载完后才去异步显示
   #  - action_type 关系的类别
   #  - submit_name 显示的提交按钮的内容
   #  - show_count 是否显示发生关系的人数
@@ -18,14 +18,14 @@ module RelationsHelper
     return if relationable.is_a? Relation
 
     count = ""
-    if opts[:show_count]
+    if !opts[:cached] && opts[:show_count]
       c = relationable.send("who_#{opts[:action_type]}").count
       count = " #{c}" if c > 0
     end
 
     content_tag(:div, class: "relation #{opts[:action_type]} #{opts[:class]}") do
-      # 用户已经登录，并且发生了关系
-      if current_user && relation = current_user.send("#{opts[:action_type]}_relation", relationable)
+      # 不需要 cached，且用户已经登录，并且发生了关系
+      if !opts[:cached] && current_user && relation = current_user.send("#{opts[:action_type]}_relation", relationable)
         content_tag(:span, class: "destroy") do
           button_to relation, method: :delete, class: opts[:button_class] do
             html = opts[:destroy_button_html] ?
@@ -50,7 +50,8 @@ module RelationsHelper
             end) # button_tag
             raw html
           end # form_tag
-        end
+        end # create span
+
       end
     end
   end
